[Reskin-450] update the filter component
